+++ tests/db_tests/test_api.py
@@ -1,11 +1,11 @@
-import os
+import logging
 import subprocess
 
 import pytest
 from rest_framework.test import APIClient
 
-from wkz import models
-from wkz.file_helper import fit_collector
+from tests.unit_tests.device.test_mount import lsusb_ready_to_be_mounted_device
+from wkz.device import mount
 
 
 @pytest.fixture
@@ -25,55 +25,68 @@ def test_stop(client):
         client.post("/stop/")
 
 
-def test_mount_device__failure(db, monkeypatch, client):
+def test_mount_device__no_device_connected(db, monkeypatch, client):
     # mock output of subprocess to prevent function from failing
     def dummy_output(dummy):
-        return "dummy-string"
+        return b"dummy-string"
 
     monkeypatch.setattr(subprocess, "check_output", dummy_output)
     res = client.post("/mount-device/")
 
     # mounting a device is barely possible in testing, thus at least assert that the endpoint returns 500
-    assert res.status_code == 500
+    assert res.status_code == 200
+    assert res.content.decode("utf8") == '"No Garmin device connected."'
+
+
+def test_mount_device__device_connected(db, monkeypatch, client, mock_mount_waiting_time):
+    from workoutizer import settings as django_settings
+
+    def task(func):
+        return func
+
+    # first mock decorator HUEY.task with dummy function
+    monkeypatch.setattr(django_settings.HUEY, "task", task)
+
+    # mock output of subprocess to prevent function from failing
+    def dummy_output(dummy):
+        return b"dummy-string-containing-Garmin"
+
+    monkeypatch.setattr(subprocess, "check_output", dummy_output)
+    res = client.post("/mount-device/")
+
+    # mounting a device is barely possible in testing, thus at least assert that the endpoint returns 500
+    assert res.status_code == 200
+    assert res.content.decode("utf8") == '"Found device, will mount and collect fit files."'
 
 
 @pytest.mark.parametrize("mock_dev", ["BLOCK", "MTP"])
-def test_mount_device__success(db, monkeypatch, tmpdir, client, mock_dev, caplog):
-    # prepare settings
-    target_dir = tmpdir.mkdir("tracks")
-    settings = models.get_settings()
-    settings.path_to_garmin_device = tmpdir  # source
-    settings.path_to_trace_dir = target_dir  # target
-    settings.save()
+def test_mount_device__success(db, monkeypatch, tmpdir, client, mock_dev, caplog, _mock_lsusb):
+    caplog.set_level(logging.DEBUG, logger="wkz.api")
 
     # mock output of subprocess ("lsusb") to prevent function from failing
-    def check_output(dummy):
-        return "dummy\nstring\nsome\ncontent\ncontaining\nGarmin"
-
-    monkeypatch.setattr(subprocess, "check_output", check_output)
+    _mock_lsusb(lsusb_ready_to_be_mounted_device)
 
     # mock output of actual mounting command (with actual gio output text)
-    path_to_device = "/some/dummy/path/to/device"
+    path_to_device = tmpdir
 
-    def mount(path):
+    def mount_cmd(path):
         return f"Mounted /dev/bus/usb/001/004 at {path_to_device}"
 
-    monkeypatch.setattr(fit_collector, "_mount_device_using_gio", mount)
-    monkeypatch.setattr(fit_collector, "_mount_device_using_pmount", mount)
+    monkeypatch.setattr(mount, "_mount_device_using_gio", mount_cmd)
+    monkeypatch.setattr(mount, "_mount_device_using_pmount", mount_cmd)
 
-    # mock output of _find_device_type
-    def _find_device_type(bus, dev):
+    # mock output of _determine_device_type
+    def _determine_device_type(path):
         if mock_dev == "MTP":
-            return ("MTP", "/dev/bus/usb/001/002")
+            return "MTP"
         elif mock_dev == "BLOCK":
-            return ("BLOCK", "/dev/sda")
+            return "BLOCK"
 
-    monkeypatch.setattr(fit_collector, "_find_device_type", _find_device_type)
+    monkeypatch.setattr(mount, "_determine_device_type", _determine_device_type)
 
-    # create directory to import the fit files from
-    fake_device_dir = os.path.join(tmpdir, "mtp:host", "Primary", "GARMIN", "Activity")
-    os.makedirs(fake_device_dir)
-
+    # mount device (no new fit files collected)
     res = client.post("/mount-device/")
-    assert f"successfully mounted device at: {path_to_device}" in caplog.text
+    assert "received POST request for mounting garmin device" in caplog.text
+    assert "found connected garmin device" in caplog.text
     assert res.status_code == 200
+    assert res.content.decode("utf8") == '"Found device, will mount and collect fit files."'