+++ tests/db_tests/test_import_activities.py
@@ -8,14 +8,14 @@ import pytz
 from wkz import models
 from wkz.best_sections.generic import activity_suitable_for_awards
 from wkz.demo import copy_demo_fit_files_to_track_dir
-from wkz.file_importer import run_importer__dask
+from wkz.io.file_importer import run_importer__dask
 from wkz.tools.utils import calc_md5
 from workoutizer import settings as django_settings
 
 
 def test_activity_data_in_db_after_import(import_one_activity):
     """
-    This test corresponds to tests.unit_tests.file_helper.test_fit_parser.test__parse_records
+    This test corresponds to tests.unit_tests.io.test_fit_parser.test__parse_records
     but checks the integration with the database and the django model fields.
     """