+++ tests/unit_tests/test_file_importer.py
@@ -3,8 +3,7 @@ from pathlib import Path
 
 import pytest
 
-from wkz.file_helper.parser import Parser
-from wkz.file_importer import (
+from wkz.io.file_importer import (
     _all_files_in_db_already,
     _check_and_parse_file,
     _convert_list_attributes_to_json,
@@ -13,6 +12,7 @@ from wkz.file_importer import (
     _parse_single_file,
     sport_naming_map,
 )
+from wkz.io.parser import Parser
 from wkz.tools.utils import calc_md5